rebase bot command
