+++ leaves-server/src/main/java/org/leavesmc/leaves/command/LeavesSubcommand.java
@@ -1,5 +1,6 @@
 package org.leavesmc.leaves.command;
 
+import org.bukkit.Location;
 import org.bukkit.command.CommandSender;
 
 import java.util.Collections;
@@ -8,7 +9,7 @@ import java.util.List;
 public interface LeavesSubcommand {
     boolean execute(CommandSender sender, String subCommand, String[] args);
 
-    default List<String> tabComplete(final CommandSender sender, final String subCommand, final String[] args) {
+    default List<String> tabComplete(final CommandSender sender, final String subCommand, final String[] args, Location location) {
         return Collections.emptyList();
     }
